# docs/api_spec.yaml
openapi: 3.1.0
info:
  title: 助盲跑 API
  description: 盲人陪跑志愿者匹配服务
  version: 1.0.0
servers:
- url: http://localhost:8081
  description: 本地开发环境
- url: http://47.114.113.171
  description: 阿里云生产环境（Swagger 已关闭，仅用于请求地址参考）
security:
- Bearer: []
...
paths:
...
  /api/orders/{id}/keep-waiting:
    put:
      tags:
      - order-controller
      operationId: keepWaiting
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
...
  /api/orders:
    post:
      tags:
      - order-controller
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/OrderResponse'
...
  /api/orders/{id}/respond:
    post:
      tags:
      - order-controller
      operationId: respondToDispatch
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DispatchRespondRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
...
  /api/orders/{id}/start:
    post:
      tags:
      - order-controller
      summary: 志愿者确认开始服务（DRIVER_ARRIVED → IN_PROGRESS）
      description: 志愿者到达后确认开始服务，由志愿者触发。前端移除盲人端"确认开始服务"按钮后，改由此接口替代。
      operationId: startService
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
        '403':
          description: 非订单参与者
        '409':
          description: 订单状态不允许此操作
  /api/orders/{id}/finish:
    post:
      tags:
      - order-controller
      operationId: finishOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
  /api/orders/{id}/en-route:
    post:
      tags:
      - order-status-controller
      operationId: driverEnRoute
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
  /api/orders/{id}/cancel:
    post:
      tags:
      - order-controller
      operationId: cancelOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
  /api/orders/{id}/arrived:
    post:
      tags:
      - order-status-controller
      operationId: driverArrived
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: object
...
  /api/orders/{id}:
    get:
      tags:
      - order-controller
      operationId: getOrder
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/OrderDetailResponse'
  /api/orders/{id}/status-logs:
    get:
      tags:
      - order-status-controller
      operationId: getStatusLogs
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderStatusLogResponse'
...
  /api/orders/mine:
    get:
      tags:
      - order-controller
      operationId: getMyOrders
      parameters:
      - name: role
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
          maximum: 100
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/PageOrderDetailResponse'
...
components:
  schemas:
...
    CreateOrderRequest:
      type: object
      properties:
        startLatitude:
          type: number
          format: double
          maximum: 90
          minimum: -90
        startLongitude:
          type: number
          format: double
          maximum: 180
          minimum: -180
        startAddress:
          type: string
        plannedStartTime:
          type: string
          format: date-time
        plannedEndTime:
          type: string
          format: date-time
        expectedDurationMinutes:
          type: integer
          format: int32
          maximum: 300
          minimum: 10
        pacePreference:
          type: string
          enum:
          - WALK_RUN
          - EASY
          - MODERATE
          - FAST
          - NO_PREFERENCE
        routePreference:
          type: string
          enum:
          - PARK_TRAIL
          - STREET
          - TRACK
          - NO_PREFERENCE
        routeNotes:
          type: string
          maxLength: 200
          minLength: 0
        hasGuideDogThisRun:
          type: boolean
        specialNotes:
          type: string
          maxLength: 200
          minLength: 0
      required:
      - plannedEndTime
      - plannedStartTime
      - startAddress
      - startLatitude
      - startLongitude
    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING_MATCH
          - PENDING_ACCEPT
          - IN_PROGRESS
          - DRIVER_EN_ROUTE
          - DRIVER_ARRIVED
          - COMPLETED
          - CANCELLED
          - REMATCHING
          - NO_VOLUNTEER
        message:
          type: string
...
    DispatchRespondRequest:
      type: object
      properties:
        action:
          type: string
          enum:
          - ACCEPT
          - DECLINE
      required:
      - action
...
    OrderDetailResponse:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING_MATCH
          - PENDING_ACCEPT
          - IN_PROGRESS
          - DRIVER_EN_ROUTE
          - DRIVER_ARRIVED
          - COMPLETED
          - CANCELLED
          - REMATCHING
          - NO_VOLUNTEER
        startAddress:
          type: string
        plannedStart:
          type: string
          format: date-time
        plannedEnd:
          type: string
          format: date-time
        volunteerPhone:
          type: string
        acceptedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        expectedDurationMinutes:
          type: integer
          format: int32
        pacePreference:
          type: string
          enum:
          - WALK_RUN
          - EASY
          - MODERATE
          - FAST
          - NO_PREFERENCE
        routePreference:
          type: string
          enum:
          - PARK_TRAIL
          - STREET
          - TRACK
          - NO_PREFERENCE
        routeNotes:
          type: string
        hasGuideDogThisRun:
          type: boolean
        specialNotes:
          type: string
        visionLevel:
          type: string
          enum:
          - TOTAL_BLIND
          - LOW_VISION
        tetherPreference:
          type: string
          enum:
          - TETHER_ROPE
          - ARM_HOLD
          - VERBAL_ONLY
        chatPreference:
          type: string
          enum:
          - PREFER_CHAT
          - PREFER_QUIET
          - NO_PREFERENCE
    OrderStatusLogResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        fromStatus:
          type: string
        toStatus:
          type: string
        changedBy:
          type: integer
          format: int64
        changedAt:
          type: string
          format: date-time
        remark:
          type: string
    PageOrderDetailResponse:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/PageableObject'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/OrderDetailResponse'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/SortObject'
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        paged:
          type: boolean
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        unpaged:
          type: boolean
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/SortObject'
    SortObject:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
...
  securitySchemes:
    Bearer:
      type: http
      name: Authorization
      scheme: bearer
      bearerFormat: JWT